Fix windows test
