+++ test/llm/ui/default/app/completion/test_completer_arguments.py
@@ -219,8 +219,10 @@ def test_file_at_prefix_path_navigation(mock_history_manager, complete_event, tm
     target = os.path.join(str(tmp_path), "be")
     doc = Document(text=f"@{target}", cursor_position=len(f"@{target}"))
     completions = list(completer.get_completions(doc, complete_event))
-    # @ completion is only_files=False, so the directory is offered.
-    assert any(c.display_text == f"beta{os.sep}" for c in completions)
+    # @ completion is only_files=False, so the directory is offered. The
+    # trailing mark is PathCompleter's own "/", not os.sep -- it does not vary
+    # by platform.
+    assert any(c.display_text == "beta/" for c in completions)
 
 
 def test_fuzzy_walk_too_many_files_falls_back_to_path_completer(
